modified printNotif() with a debug level flag system
I created a new variable uint8 debug_level. If it equals 0, it exits printNotif() without printing anything. If it equals 1, it only prints warnings and errors. If it equals 2, it prints everything. I confirmed it worked by tested all three levels with my PuTTY serial reader. The only issue is that if we set it to equal 1, it will not print the MEID or SIM ID as that is NOTIF_TYPE_EVENT.