added new print event NOTIF_TYPE_STARTUP
this new print event NOTIF_TYPE_STARTUP was added so that we can print some items that are needed for registering the device without having to print all events. (also fixed some spelling errors in the code comments)
1. prints ---BEGIN TESTS--- and ---END TESTS--- under this event type.
2. prints Device ID, SIM, MEID under this event type.

Tested:
Confirmed it works by changing the debug_level from 0u to 1u to confirm it prints everything it should and that it continues to push the data correctly to the server.