Added SDI12 and Sontek sensor capabilities
Using cellboard-SDI12 code, copied over the SDI12.c and .h files to notoriOS (called the same thing). clear_str() was copied from the strlib files from cellboard-SDI12 directly into the notoriOS SDI-12 files. debug.h was included in SDI12.c for strextract(). Sontek.c and .h files were created and sontek.c was modeled similar to the take_readings in the main.c file of cellboard-SDI12.

Power Vbat2 was assigned to be the sontek SDI12 power supply. sontek.h was added to notoriOS.h

The current set up will turn on the sontek sensor, check when it's powered on, and then take a reading. The only reading responses are the first 4 parameters available out of the 26 which may be due to the M and D commands for the sontek sensors.

The function to take sontek measurements is currently in ChickityCheckYourselfBeforeYouWreckYourself() at the top so it's the first thing that runs.

Declaring the SDI12_sensor sontek struct only worked when inside the function and not declared above it (for unknown reasons).

the sontek_t struct will need to be modified to have more of the output values wanted. the SDI12 struct will have to be modified for the metadata in SDI12.h. Will want to change output errors/number in sontek.c.