Added reading all 27 sontek parameters in SDI12.c
Change nvars for sontek to 27 (also collects ice score for ease of sending the M and D commands).

SDI12 was heavily changed to be tailored to sontek in order to loop through sending M1 & M2 along with D1 and D2 commands and to store all the readings into a sensor.nvars = 27 specifically. This was to ensure that the functions weren't repeated between sontek.c and sdi12.c. The additional commands were added to the SDI12.h file. Going forward, this will all only happen in sontek.c and more functions for the different steps of SDI12_take_measurements() can be made specifically for sontek in sontek.c to reduce the amount of loops if desired.