Adding water quality measurement capability
Water quality was originally added on the dissolved oxygen branch which was created this summer. Changes were continually made on this branch since then but changes were also made on the master branch. To ensure all master branch changes were included, a new water_quality branch was created and all the changes from the diss_oxy branch was added to it. This push includes all of those changes except the pins and top design which will be in the next push.